[WIP] [LA-372] influx reporting in gating tooling
Gating could ship this small tool to query influx db based
on env vars. When running a job, we could check for the presence
of the tool, and run it. That would output results to stdout.

On top of that, we could run the tool with parameters to have
artefacts (YAML/subunit) of the data.