RE-2217 Reflect sub build status in job result
The bump snapshots job builds new images by kicking off a new Jenkins
buid for each iamge. If those sub jobs fail, image update is skipped
for that image, but the stage shows green in the UI and the parent job
ends as successful.

This commit changes the error handling so that builds that fail show
up as failed parallel stages in blue ocean. It also fails the parent job
if any of the sub jobs fail. Note that a Jira issue is not created for
failure unless a sub job has failed 7 times consecutively.